moved location of open config button
